More reorg of mailroom types
